# [CARBONDATA-2965] Support scan performance benchmark in CarbonCli tool