… tools/cli/pom.xml
… tools/cli/src/main/java/org/apache/carbondata/tool/FileCollector.java